Add files via upload
This project utilizes web scraping techniques to programmatically extract structured data from the Worldometer website, specifically focusing on global country statistics. The process involves using the Python libraries requests to retrieve the HTML content of the specified URL and BeautifulSoup to parse the HTML structure. The objective is to systematically collect data pertaining to country details, including population figures, global population share, and land area, which is presented in a tabular format on the webpage.
'https://www.worldometers.info/geography/how-many-countries-are-there-in-the-world/'